Move restore from engine to protocol
