+++ modules/engine/src/testing/utils.spec.ts
@@ -59,8 +59,6 @@ describe(testName, () => {
   let store: Sinon.SinonStubbedInstance<MemoryStoreService>;
   let chainService: Sinon.SinonStubbedInstance<VectorChainService>;
   let messaging: Sinon.SinonStubbedInstance<MemoryMessagingService>;
-  let acquireRestoreLockStub: Sinon.SinonStub;
-  let releaseRestoreLockStub: Sinon.SinonStub;
   let withdrawRetryForTrasferIdStub: Sinon.SinonStub;
 
   // Create an EVT to post to, that can be aliased as a
@@ -92,9 +90,6 @@ describe(testName, () => {
     messaging = Sinon.createStubInstance(MemoryMessagingService);
     vector.on = on as any;
 
-    // By default acquire/release for restore succeeds
-    acquireRestoreLockStub = Sinon.stub().resolves(Result.ok(undefined));
-    releaseRestoreLockStub = Sinon.stub().resolves(Result.ok(undefined));
     withdrawRetryForTrasferIdStub = Sinon.stub(utils, "withdrawRetryForTransferId");
   });
