+++ modules/protocol/src/vector.ts
@@ -22,14 +22,20 @@ import {
   UpdateIdentifier,
 } from "@connext/vector-types";
 import { v4 as uuidV4 } from "uuid";
-import { getCreate2MultisigAddress, getRandomBytes32, delay } from "@connext/vector-utils";
+import {
+  getCreate2MultisigAddress,
+  getRandomBytes32,
+  delay,
+  getSignerAddressFromPublicIdentifier,
+  generateMerkleRoot,
+} from "@connext/vector-utils";
 import { Evt } from "evt";
 import pino from "pino";
 
-import { QueuedUpdateError } from "./errors";
+import { QueuedUpdateError, RestoreError } from "./errors";
 import { Cancellable, OtherUpdate, SelfUpdate, SerializedQueue } from "./queue";
 import { outbound, inbound, OtherUpdateResult, SelfUpdateResult } from "./sync";
-import { extractContextFromStore, persistChannel, validateParamSchema } from "./utils";
+import { extractContextFromStore, persistChannel, validateChannelSignatures, validateParamSchema } from "./utils";
 
 type EvtContainer = { [K in keyof ProtocolEventPayloadsMap]: Evt<ProtocolEventPayloadsMap[K]> };
 
@@ -611,6 +617,81 @@ export class Vector implements IVectorProtocol {
       },
     );
 
+    // response to restore messages
+    await this.messagingService.onReceiveRestoreStateMessage(
+      this.publicIdentifier,
+      async (restoreData: Result<{ chainId: number }, ProtocolError>, from: string, inbox: string) => {
+        // If it is from yourself, do nothing
+        if (from === this.publicIdentifier) {
+          return;
+        }
+        const method = "onReceiveRestoreStateMessage";
+        this.logger.debug({ method, data: restoreData.toJson(), inbox }, "Handling restore message");
+
+        // Received error from counterparty
+        if (restoreData.isError) {
+          this.logger.error(
+            { message: restoreData.getError()!.message, method },
+            "Error received from counterparty restore",
+          );
+          return;
+        }
+
+        const data = restoreData.getValue();
+        const [key] = Object.keys(data ?? []);
+        if (key !== "chainId") {
+          this.logger.error({ data }, "Message malformed");
+          return;
+        }
+
+        // Counterparty looking to initiate a restore
+        let channel: FullChannelState | undefined;
+        const sendCannotRestoreFromError = (error: Values<typeof RestoreError.reasons>, context: any = {}) => {
+          return this.messagingService.respondToRestoreStateMessage(
+            inbox,
+            Result.fail(new RestoreError(error, channel!, this.publicIdentifier, { ...context, method })),
+          );
+        };
+
+        // Get info from store to send to counterparty
+        const { chainId } = data as any;
+        try {
+          channel = await this.storeService.getChannelStateByParticipants(this.publicIdentifier, from, chainId);
+        } catch (e) {
+          return sendCannotRestoreFromError(RestoreError.reasons.CouldNotGetChannel, {
+            storeMethod: "getChannelStateByParticipants",
+            chainId,
+            identifiers: [this.publicIdentifier, from],
+          });
+        }
+        if (!channel) {
+          return sendCannotRestoreFromError(RestoreError.reasons.ChannelNotFound, { chainId });
+        }
+        let activeTransfers: FullTransferState[];
+        try {
+          activeTransfers = await this.storeService.getActiveTransfers(channel.channelAddress);
+        } catch (e) {
+          return sendCannotRestoreFromError(RestoreError.reasons.CouldNotGetActiveTransfers, {
+            storeMethod: "getActiveTransfers",
+            chainId,
+            channelAddress: channel.channelAddress,
+          });
+        }
+
+        // Send info to counterparty
+        this.logger.info(
+          {
+            method,
+            channel: channel.channelAddress,
+            nonce: channel.nonce,
+            activeTransfers: activeTransfers.map((a) => a.transferId),
+          },
+          "Sending counterparty state to sync",
+        );
+        await this.messagingService.respondToRestoreStateMessage(inbox, Result.ok({ channel, activeTransfers }));
+      },
+    );
+
     // Handle disputes
     // TODO: if this is awaited, then it may cause problems with the
     // server-node startup (double check on prod). If it is *not* awaited
@@ -802,6 +883,117 @@ export class Vector implements IVectorProtocol {
       "Method complete",
     );
     return returnVal;
+  }
+
+  public async restoreState(
+    params: ProtocolParams.Restore,
+  ): Promise<Result<FullChannelState, RestoreError | QueuedUpdateError>> {
+    const method = "restoreState";
+    const methodId = getRandomBytes32();
+    this.logger.debug({ method, methodId }, "Method start");
+    // Validate all input
+    const error = validateParamSchema(params, ProtocolParams.RestoreSchema);
+    if (error) {
+      return Result.fail(error);
+    }
+
+    // Send message to counterparty, they will grab lock and
+    // return information under lock, initiator will update channel,
+    // then send confirmation message to counterparty, who will release the lock
+    const { chainId, counterpartyIdentifier } = params;
+    const restoreDataRes = await this.messagingService.sendRestoreStateMessage(
+      Result.ok({ chainId }),
+      counterpartyIdentifier,
+      this.signer.publicIdentifier,
+    );
+    if (restoreDataRes.isError) {
+      return Result.fail(restoreDataRes.getError() as RestoreError);
+    }
+
+    const { channel, activeTransfers } = restoreDataRes.getValue() ?? ({} as any);
+
+    // Create helper to generate error
+    const generateRestoreError = (
+      error: Values<typeof RestoreError.reasons>,
+      context: any = {},
+    ): Result<FullChannelState, RestoreError> => {
+      // handle error by returning it to counterparty && returning result
+      const err = new RestoreError(error, channel, this.publicIdentifier, {
+        ...context,
+        method,
+        params,
+      });
+      return Result.fail(err);
+    };
+
+    // Verify data exists
+    if (!channel || !activeTransfers) {
+      return generateRestoreError(RestoreError.reasons.NoData);
+    }
+
+    // Verify channel address is same as calculated
+    const counterparty = getSignerAddressFromPublicIdentifier(counterpartyIdentifier);
+    const calculated = await this.chainReader.getChannelAddress(
+      channel.alice === this.signer.address ? this.signer.address : counterparty,
+      channel.bob === this.signer.address ? this.signer.address : counterparty,
+      channel.networkContext.channelFactoryAddress,
+      chainId,
+    );
+    if (calculated.isError) {
+      return generateRestoreError(RestoreError.reasons.GetChannelAddressFailed, {
+        getChannelAddressError: jsonifyError(calculated.getError()!),
+      });
+    }
+    if (calculated.getValue() !== channel.channelAddress) {
+      return generateRestoreError(RestoreError.reasons.InvalidChannelAddress, {
+        calculated: calculated.getValue(),
+      });
+    }
+
+    // Verify signatures on latest update
+    const sigRes = await validateChannelSignatures(
+      channel,
+      channel.latestUpdate.aliceSignature,
+      channel.latestUpdate.bobSignature,
+      "both",
+    );
+    if (sigRes.isError) {
+      return generateRestoreError(RestoreError.reasons.InvalidSignatures, {
+        recoveryError: sigRes.getError()!.message,
+      });
+    }
+
+    // Verify transfers match merkleRoot
+    const root = generateMerkleRoot(activeTransfers);
+    if (root !== channel.merkleRoot) {
+      return generateRestoreError(RestoreError.reasons.InvalidMerkleRoot, {
+        calculated: root,
+        merkleRoot: channel.merkleRoot,
+        activeTransfers: activeTransfers.map((t) => t.transferId),
+      });
+    }
+
+    // Verify nothing with a sync-able nonce exists in store
+    const existing = await this.getChannelState(channel.channelAddress);
+    const nonce = existing?.nonce ?? 0;
+    const diff = channel.nonce - nonce;
+    if (diff <= 1 && channel.latestUpdate.type !== UpdateType.setup) {
+      return generateRestoreError(RestoreError.reasons.SyncableState, {
+        existing: nonce,
+        toRestore: channel.nonce,
+      });
+    }
+
+    // Save channel
+    try {
+      await this.storeService.saveChannelStateAndTransfers(channel, activeTransfers);
+    } catch (e) {
+      return generateRestoreError(RestoreError.reasons.SaveChannelFailed, {
+        saveChannelStateAndTransfersError: e.message,
+      });
+    }
+
+    return Result.ok(channel);
   }
 
   ///////////////////////////////////
